acs/requirements.txt
Django==4.2.3
djangorestframework==3.14.0
channels==4.0.0
daphne==4.0.0
channels-redis==4.1.0
psycopg2-binary==2.9.6
redis==4.6.0
whitenoise
requests
Pillow
